Debugger: Fix mem lock with optimizations disabled

+++ Core/MIPS/JitCommon/JitCommon.cpp
@@ -328,8 +328,8 @@ std::vector<std::string> DisassembleRV64(const u8 *data, int size) {
 		if (len == 0) {
 			// Force align in case we're somehow unaligned.
 			len = 2 - ((uintptr_t)data & 1);
-			invalid_count += len;
-			i += len;
+			invalid_count += (int)len;
+			i +=(int) len;
 			continue;
 		}
